Use writer lock for cancel.
Fixes https://github.com/francoispqt/gojay/issues/97

By using the writer lock, you ensure that close is called only once. Closing the chan should be seen as a write operation and not a read operation. 

By using the read lock, there is a possibility that concurrent calls to cancel will cause a panic. Using a write lock fixes this.